Working loadbalancer.  Restructured packages for loadbalancer and issuer.
