+++ cmd/canis-didcomm-issuer/main.go
@@ -7,7 +7,7 @@ SPDX-License-Identifier: Apache-2.0
 package main
 
 import (
-	"github.com/scoir/canis/pkg/didcomm/cmd"
+	"github.com/scoir/canis/pkg/didcomm/issuer/cmd"
 )
 
 func main() {
