+++ docker/canis/Dockerfile
@@ -5,9 +5,8 @@ RUN apt-get install -y ca-certificates
 RUN DEBIAN_FRONTEND=noninteractive apt-get install -y tzdata
 
 COPY ./bin/canis-apiserver /usr/local/bin/canis-apiserver
-COPY ./bin/canis-scheduler /usr/local/bin/canis-scheduler
-COPY ./bin/canis-scheduler /usr/local/bin/canis-didcomm
-COPY ./bin/canis-scheduler /usr/local/bin/canis-didcomm-lb
+COPY ./bin/canis-didcomm-issuer /usr/local/bin/canis-didcomm-issuer
+COPY ./bin/canis-didcomm-lb /usr/local/bin/canis-didcomm-lb
 COPY ./bin/agent /usr/local/bin/agent
 COPY ./bin/sirius /usr/local/bin/sirius
